[None][refactor] Move cutedsl_matmul_heuristics into custom_ops/
The module is only consumed by cute_dsl_custom_ops.py; placing it
alongside its sole caller avoids a cross-sibling import and keeps
_torch/ top-level cleaner. Update all references (import, test, doc).
Fix isort formatting in cute_dsl_custom_ops.py (restore horizontal
grid style that the PR inadvertently changed to vertical).
